Move chunker from x package to a new loadfile package. Remove bulk chunker.
