+++ loadfile/chunk_test.go
@@ -13,7 +13,8 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package x
+
+package loadfile
 
 import (
 	"bufio"
